requirements.txt
gradio==4.36.1
langchain==0.2.13
langgraph==0.1.4
chromadb==0.5.5
python-docx==1.1.2
sentence-transformers
openai
unstructured[docx]
langchain-openai
pydantic==2.8.2
python-dotenv